+++ NEWS.md
@@ -1,9 +1,12 @@
-# TreeSearch 1.2.0.9001
+# TreeSearch 1.2.0.9002
+
+- New function `LengthAdded()` tests which characters contribute to taxon
+  instability, per Pol & Escapa (2009) <doi:10.1111/j.1096-0031.2009.00258.x>
 
 - New vignette on tree space mapping
 
 - Support reading characters from Excel spreadsheets in GUI
-
+  
 
 # TreeSearch 1.2.0
 
@@ -23,6 +26,7 @@
 - Improve error handling
 - Replace `throw` with `stop` in C++
 - Remove test of elapsed times, for CRAN compliance
+
 
 # TreeSearch 1.1.1